% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizBrowser-class.R
\name{.fullName,EpivizBrowser-method}
\alias{.fullName,EpivizBrowser-method}
\title{TODO: Separate class and methods}
\usage{
\S4method{.fullName}{EpivizBrowser}(x)
}
\description{
TODO: Separate class and methods
}
