% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizBrowser-class.R
\name{EpivizBrowser}
\alias{EpivizBrowser}
\title{Initialization function}
\usage{
EpivizBrowser(
  chr = "chr1",
  start = 1,
  end = 1e+05,
  genome,
  tracks = list(),
  ...
)
}
\arguments{
\item{chr}{chromosome region to intitialize}

\item{start}{start region}

\item{end}{end region}
}
\description{
Initialization function
}
